+++ src/renderer/components/Palettes/Palettes.js
@@ -1,108 +1,72 @@
 import React from 'react'
+import { remote } from 'electron'
 import styled from 'styled-components'
-import Palette from './Palette'
+import NoPalettes from './NoPalettes'
+import TitleList from './TitleList'
+import Cube from './Cube'
+import IconRow from './IconRow'
+import ColorRow from './ColorRow'
+import copyToClipboard from '../../lib/copyToClipboard'
+import prompt from 'common/electronPopup'
 
-const Container = styled.div`
-  height: calc(100vh - 55px);
+export const Container = styled.div`
+  height: calc(100vh - 50px);
   display: grid;
   grid-template-columns: 200px 1fr;
   padding: 0.5rem;
-`
-
-const List = styled.div`
-  display: grid;
-  grid-template-rows: 5px repeat(auto-fit, 30px);
-`
-
-const ListItem = styled.p`
-  width: 170px;
-  height: 30px;
-  display: grid;
-  align-items: center;
-  font-size: 1.25rem;
-  margin: 0;
-  padding: 0rem 1rem;
-  cursor: pointer;
-  user-select: none;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  text-align: left;
-  white-space: nowrap;
-  background: ${props => (props.selected ? props.theme.black : 'transparent')};
-  color: ${props => (props.selected ? 'white' : props.theme.black)};
 `
 
 const Display = styled.div`
   height: 95%;
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
   align-items: center;
-  p {
-    width: 75%;
-    font-family: 'Oswald';
-    font-size: 3rem;
-    text-align: center;
+  .top {
+    width: 100%;
+    display: grid;
+    grid-template-columns: 1fr 2fr 1.5fr;
     margin: 0;
     padding: 0;
     background: white;
     border: ${props => props.theme.border};
+    p {
+      justify-self: center;
+      font-family: 'Oswald';
+      font-size: 2.5rem;
+      text-align: center;
+      margin: 0;
+      padding: 0;
+    }
+    .actions {
+      align-self: flex-end;
+      justify-self: flex-end;
+      button {
+        background: white;
+        color: ${props => props.theme.black};
+        border: none;
+        cursor: pointer;
+        transition: all 0.25s;
+        &:hover {
+          color: dodgerblue;
+        }
+      }
+    }
   }
   .grid {
+    width: 100%;
+    height: 350px;
+    overflow-y: auto;
     display: grid;
-  }
-  button {
-    width: 100px;
-    background: white;
-    border: ${props => props.theme.border};
-    padding: 0.5rem 1rem;
-    margin-right: 0.75rem;
-    transition: all 0.25s;
-    user-select: none;
-    &:hover {
-      background: ${props => props.theme.black};
-      color: white;
-    }
-  }
-`
-
-const NoPalettes = styled.div`
-  display: grid;
-  grid-template-rows: auto auto;
-  justify-items: center;
-  align-items: center;
-  p {
-    background: white;
-    color: ${props => props.theme.black};
-    font-size: 3rem;
-    font-family: 'Oswald';
-    padding: 1rem;
-    margin: 0;
-    border: ${props => props.theme.border};
-  }
-  button {
-    width: 100px;
-    background: white;
-    color: ${props => props.theme.black};
-    padding: 0.5rem 1rem;
-    user-select: none;
-    transition: all 0.25s;
-    &:hover {
-      background: ${props => props.theme.black};
-      color: white;
-    }
+    grid-template-rows: ${props => `repeat(${props.length}, 35px)`};
+    grid-gap: 5px;
+    justify-items: center;
+    margin-top: 1rem;
   }
 `
 
 export default class Palettes extends React.Component {
   state = {
     index: 0
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (prevProps.palettes.length !== this.props.palettes.length) {
-      this.setState({ index: 0 })
-    }
   }
 
   handleDelete = (i, name) => {
@@ -113,38 +77,55 @@ export default class Palettes extends React.Component {
 
   setIndex = index => this.setState({ index })
 
-  render() {
+  copyAsVariables = () => {
     const { index } = this.state
     const { palettes } = this.props
+    let colors = palettes[index].colors
+    let str = ``
+    colors.forEach(c => {
+      if (c.name) {
+        str += `${c.name}: ${c.color};\n`
+      }
+    })
+    copyToClipboard(str)
+    const options = {
+      height: 200,
+      title: 'Success',
+      label: 'Copied Text',
+      message:
+        'Your color names and values have been copied to your clipboard in css/scss variable style.',
+      ok: 'I Get It',
+      type: 'alert'
+    }
+    prompt(options, remote.getCurrentWindow())
+  }
+
+  render() {
+    const { index } = this.state
+    const { setMode, palettes } = this.props
     if (!palettes.length) {
-      return (
-        <NoPalettes>
-          <p>No Saved Palettes</p>
-          <button onClick={() => this.props.setMode(0)}>Back</button>
-        </NoPalettes>
-      )
+      return <NoPalettes setMode={setMode} />
     } else {
       const { name, colors } = palettes[index]
       return (
         <Container>
-          <List>
-            <span />
-            {palettes.map((p, i) => (
-              <ListItem key={p.name} selected={index === i} onClick={() => this.setIndex(i)}>
-                {p.name}
-              </ListItem>
-            ))}
-          </List>
-          <Display>
-            <p>{name}</p>
-            <div className="grid">
-              <Palette colors={colors} onSwatchClick={() => {}} onContextMenu={() => {}} />
+          <TitleList palettes={palettes} index={index} setIndex={this.setIndex} />
+          <Display length={colors.filter(c => !c.clean).length}>
+            <div className="top">
+              <Cube colors={colors} />
+              <p>{name}</p>
+              <IconRow
+                setMode={() => this.props.setMode(0)}
+                copyAsVariables={this.copyAsVariables}
+                loadPalette={() => this.props.loadPalette(palettes[index])}
+                handleDelete={() => this.handleDelete(index, name)}
+              />
             </div>
-
-            <div>
-              <button onClick={() => this.props.loadPalette(palettes[index])}>Load Palette</button>
-              <button onClick={() => this.handleDelete(index, name)}>Delete Palette</button>
-              <button onClick={() => this.props.setMode(0)}>Back</button>
+            <div className="grid">
+              {colors.map((c, i) => {
+                if (c.clean) return null
+                return <ColorRow key={i} color={c} />
+              })}
             </div>
           </Display>
         </Container>