+++ src/renderer/lib/getContrastColor.js
@@ -2,5 +2,5 @@ import getHSLParts from './getHSLParts'
 
 export default color => {
   const [_, h, s, l, a] = getHSLParts(color)
-  return l >= 50 ? '#000000' : '#FFFFFF'
+  return l >= 55 ? '#333333' : '#EEEEEE'
 }